refactor(finance): Moved out `BoundPercent`
